# Conflicts:
#	Cargo.lock
#	Cargo.toml
#	builder/Cargo.toml
#	builder/src/lib.rs
#	contract-bindings-ethers/src/permissioned_stake_table.rs
#	contracts/rust/adapter/Cargo.toml
#	contracts/rust/adapter/src/bindings/esptoken.rs
#	contracts/rust/adapter/src/bindings/mod.rs
#	contracts/rust/adapter/src/bindings/permissionedstaketable.rs
#	contracts/rust/adapter/src/bindings/staketable.rs
#	contracts/rust/adapter/src/light_client.rs
#	contracts/rust/adapter/src/stake_table.rs
#	contracts/rust/diff-test/Cargo.toml
#	contracts/src/StakeTable.sol
#	contracts/test/StakeTable.t.sol
#	contracts/test/mocks/StakeTableMock.sol
#	docker-compose.yaml
#	hotshot-builder-core/src/builder_state.rs
#	hotshot-state-prover/src/service.rs
#	hotshot-types/src/network.rs
#	hotshot-types/src/vote.rs
#	hotshot/src/traits/election/randomized_committee.rs
#	hotshot/src/traits/election/static_committee.rs
#	hotshot/src/traits/election/two_static_committees.rs
#	justfile
#	marketplace-builder/Cargo.toml
#	marketplace-builder/src/builder.rs
#	process-compose.yaml
#	sequencer-sqlite/Cargo.lock
#	sequencer/src/bin/deploy.rs
#	sequencer/src/bin/espresso-dev-node.rs
#	sequencer/src/bin/orchestrator.rs
#	sequencer/src/bin/update-permissioned-stake-table.rs
#	sequencer/src/lib.rs
#	types/src/v0/impls/header.rs
#	types/src/v0/impls/l1.rs
#	types/src/v0/impls/stake_table.rs
#	types/src/v0/impls/state.rs
#	types/src/v0/v0_1/l1.rs
#	types/src/v0/v0_3/chain_config.rs
#	types/src/v0/v0_3/stake_table.rs
#	utils/src/deployer.rs
#	utils/src/stake_table.rs

+++ contracts/rust/adapter/src/light_client.rs
@@ -45,8 +45,8 @@ impl From<LightClient::finalizedStateReturn> for LightClientStateSol {
 impl<F: PrimeField> From<LightClientStateSol> for GenericLightClientState<F> {
     fn from(v: LightClientStateSol) -> Self {
         Self {
-            view_number: v.viewNum as usize,
-            block_height: v.blockHeight as usize,
+            view_number: v.viewNum,
+            block_height: v.blockHeight,
             block_comm_root: u256_to_field(v.blockCommRoot),
         }
     }
